Restructure C++ into multiple files
