+++ .gitignore
@@ -2,3 +2,4 @@ deps/build.log
 Artifacts.toml
 Manifest.toml
 *.fits
+*.so
